requirements.txt
python-dotenv==1.0.0
telethon==1.27.0
pyflakes==3.0.1
flake8==6.0.0
flask==2.2.3
marshmallow==3.19.0
celery==5.2.7
pymongo==4.3.3